refactor(nav-drawer): Barrel file renaming
