+++ projects/igniteui-angular/src/lib/navigation-drawer/public_api.ts
