+++ projects/igniteui-angular/src/public_api.ts
@@ -81,7 +81,7 @@ export * from './lib/input-group/public_api';
 export * from './lib/list/public_api';
 export * from './lib/expansion-panel/public_api';
 export * from './lib/navbar/navbar.component';
-export * from './lib/navigation-drawer/index';
+export * from './lib/navigation-drawer/public_api';
 export * from './lib/paginator/paginator.component';
 export * from './lib/progressbar/progressbar.component';
 export * from './lib/radio/radio.component';
